# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OfferComparisonOut": {
        "additionalProperties": false,
        "properties": {
          "scenarioGroups": {
            "items": {
              "$ref": "#/components/schemas/ScenarioGroupOut"
            },
            "title": "Scenariogroups",
            "type": "array"
          },
          "sortKey": {
            "enum": [
              "price",
              "time",
              "score"
            ],
            "title": "Sortkey",
            "type": "string"
          }
        },
        "required": [
          "sortKey",
          "scenarioGroups"
        ],
        "title": "OfferComparisonOut",
        "type": "object"
      },
      "OfferOut": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "notSelectableReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notselectablereason"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "selectable": {
            "title": "Selectable",
            "type": "boolean"
          },
          "specialist": {
            "$ref": "#/components/schemas/SpecialistPublicOut"
          },
          "status": {
            "$ref": "#/components/schemas/OfferStatus"
          },
          "version": {
            "$ref": "#/components/schemas/OfferVersionOut"
          }
        },
        "required": [
          "id",
          "status",
          "revision",
          "specialist",
          "version",
          "selectable"
        ],
        "title": "OfferOut",
        "type": "object"
      },
      "OfferStatus": {
        "enum": [
          "active",
          "withdrawn",
          "selected",
          "expired",
          "unselectable"
        ],
        "title": "OfferStatus",
        "type": "string"
      },
      "OfferType": {
        "enum": [
          "fixed",
          "conditional",
          "diagnostic"
        ],
        "title": "OfferType",
        "type": "string"
      },
      "OfferVersionOut": {
        "additionalProperties": false,
        "properties": {
          "conditionsNote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conditionsnote"
          },
          "diagnosticFeeCredited": {
            "title": "Diagnosticfeecredited",
            "type": "boolean"
          },
          "diagnosticFeeToman": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnosticfeetoman"
          },
          "diagnosticScope": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnosticscope"
          },
          "estimatedMinutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimatedminutes"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "lines": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Lines",
            "type": "array"
          },
          "offerType": {
            "$ref": "#/components/schemas/OfferType"
          },
          "scenarios": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Scenarios",
            "type": "array"
          },
          "scheduledAt": {
            "format": "date-time",
            "title": "Scheduledat",
            "type": "string"
          },
          "specialistPayableToman": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specialistpayabletoman"
          },
          "totalToman": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Totaltoman"
          },
          "validUntil": {
            "format": "date-time",
            "title": "Validuntil",
            "type": "string"
          },
          "versionNumber": {
            "title": "Versionnumber",
            "type": "integer"
          },
          "warrantyNote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warrantynote"
          }
        },
        "required": [
          "id",
          "versionNumber",
          "offerType",
          "lines",
          "scenarios",
          "totalToman",
          "specialistPayableToman",
          "warrantyNote",
          "conditionsNote",
          "diagnosticScope",
          "diagnosticFeeToman",
          "diagnosticFeeCredited",
          "estimatedMinutes",
          "scheduledAt",
          "validUntil"
        ],
        "title": "OfferVersionOut",
        "type": "object"
      },
...
      "RefundInput": {
        "additionalProperties": false,
        "properties": {
          "note": {
            "anyOf": [
              {
                "maxLength": 600,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "title": "RefundInput",
        "type": "object"
      },
      "RefundOut": {
        "additionalProperties": false,
        "properties": {
          "amountToman": {
            "title": "Amounttoman",
            "type": "integer"
          },
          "customerNote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customernote"
          },
          "decisionNote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decisionnote"
          },
          "findings": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Findings",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "isSample": {
            "default": true,
            "title": "Issample",
            "type": "boolean"
          },
          "policyMode": {
            "$ref": "#/components/schemas/RefundPolicyMode"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "reviewDeadline": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewdeadline"
          },
          "status": {
            "$ref": "#/components/schemas/RefundStatus"
          },
          "transferAttempts": {
            "title": "Transferattempts",
            "type": "integer"
          },
          "transferredAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transferredat"
          }
        },
        "required": [
          "id",
          "status",
          "amountToman",
          "reason",
          "policyMode",
          "customerNote",
          "decisionNote",
          "reviewDeadline",
          "transferredAt",
          "transferAttempts",
          "findings"
        ],
        "title": "RefundOut",
        "type": "object"
      },
      "RefundPolicyMode": {
        "enum": [
          "bootstrap",
          "reference_based"
        ],
        "title": "RefundPolicyMode",
        "type": "string"
      },
      "RefundStatus": {
        "enum": [
          "requested",
          "reviewing",
          "needs_review",
          "approved",
          "rejected",
          "transfer_pending",
          "paid",
          "transfer_failed"
        ],
        "title": "RefundStatus",
        "type": "string"
      },
...
      "ScenarioGroupOut": {
        "additionalProperties": false,
        "properties": {
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "offers": {
            "items": {
              "$ref": "#/components/schemas/OfferOut"
            },
            "title": "Offers",
            "type": "array"
          },
          "scenarioCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenariocode"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "scenarioCode",
          "title",
          "offers"
        ],
        "title": "ScenarioGroupOut",
        "type": "object"
      },
      "ScenarioInput": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 60,
            "title": "Code",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 400,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "title": {
            "maxLength": 200,
            "title": "Title",
            "type": "string"
          },
          "totalToman": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Totaltoman"
          }
        },
        "required": [
          "code",
          "title"
        ],
        "title": "ScenarioInput",
        "type": "object"
      },
...
      "SpecialistPublicOut": {
        "additionalProperties": false,
        "description": "What a customer may see about a bidder. No contact details before acceptance.",
        "properties": {
          "assessableCases": {
            "default": 0,
            "title": "Assessablecases",
            "type": "integer"
          },
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "district": {
            "title": "District",
            "type": "string"
          },
          "hasEnoughHistory": {
            "default": false,
            "title": "Hasenoughhistory",
            "type": "boolean"
          },
          "invoiceOverpriceCases": {
            "default": 0,
            "title": "Invoiceoverpricecases",
            "type": "integer"
          },
          "negativeCases": {
            "default": 0,
            "title": "Negativecases",
            "type": "integer"
          },
          "satisfactionAverage": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Satisfactionaverage"
          },
          "satisfactionResponses": {
            "default": 0,
            "title": "Satisfactionresponses",
            "type": "integer"
          },
          "score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "shopName": {
            "title": "Shopname",
            "type": "string"
          },
          "specialistId": {
            "format": "uuid",
            "title": "Specialistid",
            "type": "string"
          },
          "unjustifiedIncreaseCases": {
            "default": 0,
            "title": "Unjustifiedincreasecases",
            "type": "integer"
          }
        },
        "required": [
          "specialistId",
          "displayName",
          "shopName",
          "district"
        ],
        "title": "SpecialistPublicOut",
        "type": "object"
      },
...
      "SubmitOfferInput": {
        "additionalProperties": false,
        "properties": {
          "conditionsNote": {
            "anyOf": [
              {
                "maxLength": 400,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conditionsnote"
          },
          "diagnosticFeeCredited": {
            "default": false,
            "title": "Diagnosticfeecredited",
            "type": "boolean"
          },
          "diagnosticFeeToman": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnosticfeetoman"
          },
          "diagnosticScope": {
            "anyOf": [
              {
                "maxLength": 400,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnosticscope"
          },
          "estimatedMinutes": {
            "anyOf": [
              {
                "maximum": 100000.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimatedminutes"
          },
          "expectedRevision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expectedrevision"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/LineItem"
            },
            "maxItems": 40,
            "title": "Lines",
            "type": "array"
          },
          "offerType": {
            "$ref": "#/components/schemas/OfferType"
          },
          "scenarios": {
            "items": {
              "$ref": "#/components/schemas/ScenarioInput"
            },
            "maxItems": 8,
            "title": "Scenarios",
            "type": "array"
          },
          "scheduledAt": {
            "format": "date-time",
            "title": "Scheduledat",
            "type": "string"
          },
          "validUntil": {
            "format": "date-time",
            "title": "Validuntil",
            "type": "string"
          },
          "warrantyNote": {
            "anyOf": [
              {
                "maxLength": 400,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warrantynote"
          }
        },
        "required": [
          "offerType",
          "scheduledAt",
          "validUntil"
        ],
        "title": "SubmitOfferInput",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "API نمایشی «ترب تعمیر». دادهٔ نمونه، پرداخت شبیه‌سازی‌شده و پاسخ AI نمایشی در این نصب صریحاً برچسب‌گذاری می‌شوند.",
    "title": "ترب تعمیر — API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/offers/{offer_id}/withdraw": {
      "post": {
        "operationId": "withdraw_offer_api_offers__offer_id__withdraw_post",
        "parameters": [
          {
            "in": "path",
            "name": "offer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Offer Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "expected_revision",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Expected Revision"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Withdraw Offer",
        "tags": [
          "offers"
        ]
      }
    },
...
    "/api/refunds/{refund_id}/reopen": {
      "post": {
        "description": "One free re-review with new evidence, within 7 days of a rejection.",
        "operationId": "reopen_api_refunds__refund_id__reopen_post",
        "parameters": [
          {
            "in": "path",
            "name": "refund_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Refund Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefundInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefundOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reopen",
        "tags": [
          "refunds"
        ]
      }
    },
...
    "/api/requests/{request_id}/offers": {
      "get": {
        "description": "The customer's comparison, grouped per shared scenario and ordered deterministically.",
        "operationId": "compare_offers_api_requests__request_id__offers_get",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "price",
              "enum": [
                "price",
                "time",
                "score"
              ],
              "title": "Sort",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferComparisonOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compare Offers",
        "tags": [
          "offers"
        ]
      },
      "post": {
        "description": "Price is free. Only structural defects block submission.",
        "operationId": "submit_offer_api_requests__request_id__offers_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitOfferInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Offer",
        "tags": [
          "offers"
        ]
      }
    },
...
    "/api/requests/{request_id}/refund": {
      "get": {
        "operationId": "get_refund_api_requests__request_id__refund_get",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/RefundOut"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Refund Api Requests  Request Id  Refund Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Refund",
        "tags": [
          "refunds"
        ]
      },
      "post": {
        "description": "Free to ask for, independent of the chat quota, one per payment.",
        "operationId": "request_refund_api_requests__request_id__refund_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefundInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefundOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request Refund",
        "tags": [
          "refunds"
        ]
      }
    },
...
  }
}